fix consultation range creation

+++ src/qt/daoconsultationcreate.cpp
@@ -175,7 +175,7 @@ void DaoConsultationCreate::onCreate()
         showWarning(tr("Choose a smaller number of answers."));
         return;
     }
-    else if(fRange && !(nMin >= 0 && nMax < VoteFlags::VOTE_ABSTAIN && nMax > nMin))
+    else if(fRange && !(nMin >= 0 && nMax < (uint64_t)VoteFlags::VOTE_ABSTAIN && nMax > nMin))
     {
         showWarning(tr("Wrong range!"));
         return;
